Refactor package_config into separate modules
- Extract PACKAGE_REGISTRY to package_registry.py
- Move helper functions to package_utils.py
- Keep only VersionStrategy and PackageConfig in package_config.py
- Update imports in doc_gen4.py and doc_parser.py
